+++ libs/sqlite/src/com/biffweb/sqlite/impl/system.clj
@@ -4,13 +4,7 @@
             [com.biffweb.sqlite.impl.kv :as kv]
             [com.biffweb.sqlite.impl.litestream :as litestream]
             [com.biffweb.sqlite.impl.pool :as pool]
-            [com.biffweb.sqlite.impl.sqldef :as impl.sqldef]
-            [next.jdbc :as jdbc]))
-
-(defn- wrap-db-snapshot [f]
-  (fn [ctx]
-    (jdbc/with-transaction [tx (:biff.sqlite/read-pool ctx)]
-      (f (assoc ctx :biff.sqlite/read-pool tx)))))
+            [com.biffweb.sqlite.impl.sqldef :as impl.sqldef]))
 
 (defn- start
   [ctx]
@@ -53,10 +47,6 @@
 
    :biff.core/init
    (fn [modules-var]
-     {:biff.core/kv-get           kv/get-value
-      :biff.core/kv-list          kv/list-keys
-      :biff.core/kv-set           kv/set-value
-      :biff.core/wrap-db-snapshot wrap-db-snapshot
-      :biff.sqlite/columns        (into {}
-                                        (mapcat :biff.sqlite/columns)
-                                        @modules-var)})})
+     {:biff.sqlite/columns (into {}
+                                 (mapcat :biff.sqlite/columns)
+                                 @modules-var)})})