+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 Unreleased:
 
-- biff.datastar: scope :biff.datastar/tab-id by user ID
+- biff.datastar: scope `:biff.datastar/tab-id` by user ID
+- biff.sqlite, biff.xtdb: set the biff.core DB functions via the same module
+  that sets the connections/node.
 
 **Breaking** [2.0.0-rc23](RELEASES/2.0.0-rc23.md) (2026-08-29):